add central config to disable pipeline steps

+++ webhook_handler/pipeline.py
@@ -66,25 +66,29 @@ class Pipeline:
         self._config.setup_pr_log_dir(self._pr_data.id)
         configure_logger(self._config.pr_log_dir, self._execution_id)
         self.logger = logging.getLogger()
-        helpers.remove_dir(Path(self._config.cloned_repo_dir))
+        if self._config.execute_teardown:
+            helpers.remove_dir(Path(self._config.cloned_repo_dir))
 
     def _teardown(self) -> None:
         """
         Cleans state of directory after completion.
         """
 
-        helpers.remove_dir(Path(self._config.cloned_repo_dir), log_success=True)
-        image_tag = self._pr_data.image_tag
-        try:
-            client = docker.from_env()
-            client.images.remove(image=f"{image_tag}:latest", force=True)
-            self.logger.success(f"Removed Docker image '{image_tag}'")
-        except ImageNotFound:
-            self.logger.error(f"Tried to remove image '{image_tag}', but it was not found")
-        except Exception as e:
-            self.logger.error(f"Failed to remove Docker image '{image_tag}': {e}")
-        with self.executed_tests.open("a", encoding='utf-8') as f:
-            f.write(f"{self._execution_id}\n")
+        if self._config.execute_teardown:
+            helpers.remove_dir(Path(self._config.cloned_repo_dir), log_success=True)
+            image_tag = self._pr_data.image_tag
+            try:
+                client = docker.from_env()
+                client.images.remove(image=f"{image_tag}:latest", force=True)
+                self.logger.success(f"Removed Docker image '{image_tag}'")
+            except ImageNotFound:
+                self.logger.error(f"Tried to remove image '{image_tag}', but it was not found")
+            except Exception as e:
+                self.logger.error(f"Failed to remove Docker image '{image_tag}': {e}")
+            with self.executed_tests.open("a", encoding='utf-8') as f:
+                f.write(f"{self._execution_id}\n")
+        else:
+            self.logger.warning("Teardown is disabled")
 
         self._gh_api = None
         self._issue_statement = None
@@ -248,6 +252,8 @@ class Pipeline:
         pdf_name, pdf_content = "", b""
         if self._config.fetch_pdf:
             pdf_name, pdf_content = self._pr_diff_ctx.get_issue_pdf(self._pdf_candidate, self._pr_data.head_commit)
+        else:
+            self.logger.warning("PDF fetching is disabled")
 
         # 5. Clone repository locally
         if not Path(self._config.cloned_repo_dir).exists():
@@ -260,16 +266,21 @@ class Pipeline:
         code_sliced = self._cst_builder.slice_code_file()
 
         # 7. Fetch test file for injection
-        try:
-            test_filename, test_file_content, test_file_content_sliced = test_injection.get_candidate_test_file(
-                self._config.parse_language,
-                self._pr_data.base_commit,
-                self._pr_diff_ctx.golden_code_patch,
-                self._config.cloned_repo_dir
-            )
-        except:
-            self.logger.critical("Failed to determine test file for injection")
-            raise ExecutionError("Failed to determine test file for injection")
+        test_filename = self._config.inject_in_file
+        if not test_filename:
+            try:
+                test_filename, test_file_content, test_file_content_sliced = test_injection.get_candidate_test_file(
+                    self._config.parse_language,
+                    self._pr_data.base_commit,
+                    self._pr_diff_ctx.golden_code_patch,
+                    self._config.cloned_repo_dir
+                )
+            except:
+                self.logger.critical("Failed to determine test file for injection")
+                raise ExecutionError("Failed to determine test file for injection")
+        else:
+            self.logger.warning(f"Custom test file {test_filename} is defined")
+            test_file_content = test_file_content_sliced = ""
 
         # 8. Fetch packages and imports
         try:
